# api_spec.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "FinZen API",
                 "description":  "AI-powered micro-investment platform for beginners",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/api/gamification/gamification/profile":  {
                                                                 "get":  {
                                                                             "tags":  [
                                                                                          "Gamification",
                                                                                          "Gamification"
                                                                                      ],
                                                                             "summary":  "Get Profile",
                                                                             "operationId":  "get_profile_api_gamification_gamification_profile_get",
                                                                             "parameters":  [
                                                                                                {
                                                                                                    "name":  "authorization",
                                                                                                    "in":  "header",
                                                                                                    "required":  false,
                                                                                                    "schema":  {
                                                                                                                   "anyOf":  [
                                                                                                                                 {
                                                                                                                                     "type":  "string"
                                                                                                                                 },
                                                                                                                                 {
                                                                                                                                     "type":  "null"
                                                                                                                                 }
                                                                                                                             ],
                                                                                                                   "title":  "Authorization"
                                                                                                               }
                                                                                                }
                                                                                            ],
                                                                             "responses":  {
                                                                                               "200":  {
                                                                                                           "description":  "Successful Response",
                                                                                                           "content":  {
                                                                                                                           "application/json":  {
                                                                                                                                                    "schema":  {
                                                                                                                                                                   "$ref":  "#/components/schemas/GamificationProfileResponse"
                                                                                                                                                               }
                                                                                                                                                }
                                                                                                                       }
                                                                                                       },
                                                                                               "422":  {
                                                                                                           "description":  "Validation Error",
                                                                                                           "content":  {
                                                                                                                           "application/json":  {
                                                                                                                                                    "schema":  {
                                                                                                                                                                   "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                               }
                                                                                                                                                }
                                                                                                                       }
                                                                                                       }
                                                                                           }
                                                                         }
                                                             },
                  "/api/gamification/gamification/leaderboard":  {
                                                                     "get":  {
                                                                                 "tags":  [
                                                                                              "Gamification",
                                                                                              "Gamification"
                                                                                          ],
                                                                                 "summary":  "Get Leaderboard",
                                                                                 "operationId":  "get_leaderboard_api_gamification_gamification_leaderboard_get",
                                                                                 "parameters":  [
                                                                                                    {
                                                                                                        "name":  "period",
                                                                                                        "in":  "query",
                                                                                                        "required":  false,
                                                                                                        "schema":  {
                                                                                                                       "type":  "string",
                                                                                                                       "pattern":  "^(weekly|monthly|all)$",
                                                                                                                       "default":  "all",
                                                                                                                       "title":  "Period"
                                                                                                                   }
                                                                                                    },
                                                                                                    {
                                                                                                        "name":  "limit",
                                                                                                        "in":  "query",
                                                                                                        "required":  false,
                                                                                                        "schema":  {
                                                                                                                       "type":  "integer",
                                                                                                                       "maximum":  100,
                                                                                                                       "minimum":  1,
                                                                                                                       "default":  10,
                                                                                                                       "title":  "Limit"
                                                                                                                   }
                                                                                                    },
                                                                                                    {
                                                                                                        "name":  "authorization",
                                                                                                        "in":  "header",
                                                                                                        "required":  false,
                                                                                                        "schema":  {
                                                                                                                       "anyOf":  [
                                                                                                                                     {
                                                                                                                                         "type":  "string"
                                                                                                                                     },
                                                                                                                                     {
                                                                                                                                         "type":  "null"
                                                                                                                                     }
                                                                                                                                 ],
                                                                                                                       "title":  "Authorization"
                                                                                                                   }
                                                                                                    }
                                                                                                ],
                                                                                 "responses":  {
                                                                                                   "200":  {
                                                                                                               "description":  "Successful Response",
                                                                                                               "content":  {
                                                                                                                               "application/json":  {
                                                                                                                                                        "schema":  {
                                                                                                                                                                       "$ref":  "#/components/schemas/LeaderboardResponse"
                                                                                                                                                                   }
                                                                                                                                                    }
                                                                                                                           }
                                                                                                           },
                                                                                                   "422":  {
                                                                                                               "description":  "Validation Error",
                                                                                                               "content":  {
                                                                                                                               "application/json":  {
                                                                                                                                                        "schema":  {
                                                                                                                                                                       "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                                   }
                                                                                                                                                    }
                                                                                                                           }
                                                                                                           }
                                                                                               }
                                                                             }
                                                                 },
                  "/api/gamification/gamification/daily-reward":  {
                                                                      "post":  {
                                                                                   "tags":  [
                                                                                                "Gamification",
                                                                                                "Gamification"
                                                                                            ],
                                                                                   "summary":  "Claim Daily Reward",
                                                                                   "operationId":  "claim_daily_reward_api_gamification_gamification_daily_reward_post",
                                                                                   "parameters":  [
                                                                                                      {
                                                                                                          "name":  "authorization",
                                                                                                          "in":  "header",
                                                                                                          "required":  false,
                                                                                                          "schema":  {
                                                                                                                         "anyOf":  [
                                                                                                                                       {
                                                                                                                                           "type":  "string"
                                                                                                                                       },
                                                                                                                                       {
                                                                                                                                           "type":  "null"
                                                                                                                                       }
                                                                                                                                   ],
                                                                                                                         "title":  "Authorization"
                                                                                                                     }
                                                                                                      }
                                                                                                  ],
                                                                                   "responses":  {
                                                                                                     "200":  {
                                                                                                                 "description":  "Successful Response",
                                                                                                                 "content":  {
                                                                                                                                 "application/json":  {
                                                                                                                                                          "schema":  {
                                                                                                                                                                         "$ref":  "#/components/schemas/DailyRewardResponse"
                                                                                                                                                                     }
                                                                                                                                                      }
                                                                                                                             }
                                                                                                             },
                                                                                                     "422":  {
                                                                                                                 "description":  "Validation Error",
                                                                                                                 "content":  {
                                                                                                                                 "application/json":  {
                                                                                                                                                          "schema":  {
                                                                                                                                                                         "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                                     }
                                                                                                                                                      }
                                                                                                                             }
                                                                                                             }
                                                                                                 }
                                                                               }
                                                                  },
...
              },
    "components":  {
                       "schemas":  {
                                       "DailyRewardResponse":  {
                                                                   "properties":  {
                                                                                      "reward":  {
                                                                                                     "type":  "integer",
                                                                                                     "title":  "Reward"
                                                                                                 },
                                                                                      "claimed":  {
                                                                                                      "type":  "boolean",
                                                                                                      "title":  "Claimed"
                                                                                                  }
                                                                                  },
                                                                   "type":  "object",
                                                                   "required":  [
                                                                                    "reward",
                                                                                    "claimed"
                                                                                ],
                                                                   "title":  "DailyRewardResponse"
                                                               },
...
                                       "GamificationProfileResponse":  {
                                                                           "properties":  {
                                                                                              "level":  {
                                                                                                            "type":  "integer",
                                                                                                            "title":  "Level"
                                                                                                        },
                                                                                              "points":  {
                                                                                                             "type":  "integer",
                                                                                                             "title":  "Points"
                                                                                                         },
                                                                                              "badges":  {
                                                                                                             "items":  {
                                                                                                                           "type":  "string"
                                                                                                                       },
                                                                                                             "type":  "array",
                                                                                                             "title":  "Badges"
                                                                                                         },
                                                                                              "streak":  {
                                                                                                             "type":  "integer",
                                                                                                             "title":  "Streak"
                                                                                                         }
                                                                                          },
                                                                           "type":  "object",
                                                                           "required":  [
                                                                                            "level",
                                                                                            "points",
                                                                                            "badges",
                                                                                            "streak"
                                                                                        ],
                                                                           "title":  "GamificationProfileResponse"
                                                                       },
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
                                       "LeaderboardEntry":  {
                                                                "properties":  {
                                                                                   "user":  {
                                                                                                "type":  "string",
                                                                                                "title":  "User"
                                                                                            },
                                                                                   "points":  {
                                                                                                  "type":  "integer",
                                                                                                  "title":  "Points"
                                                                                              }
                                                                               },
                                                                "type":  "object",
                                                                "required":  [
                                                                                 "user",
                                                                                 "points"
                                                                             ],
                                                                "title":  "LeaderboardEntry"
                                                            },
                                       "LeaderboardResponse":  {
                                                                   "properties":  {
                                                                                      "leaderboard":  {
                                                                                                          "items":  {
                                                                                                                        "$ref":  "#/components/schemas/LeaderboardEntry"
                                                                                                                    },
                                                                                                          "type":  "array",
                                                                                                          "title":  "Leaderboard"
                                                                                                      }
                                                                                  },
                                                                   "type":  "object",
                                                                   "required":  [
                                                                                    "leaderboard"
                                                                                ],
                                                                   "title":  "LeaderboardResponse"
                                                               },
...
                                   }
                   }
}